INTERNAL: Rename stats files to stats_prefix
