+++ Makefile.am
@@ -79,7 +79,8 @@ memcached_SOURCES = \
                     memcached.c\
                     memcached.h \
                     sasl_defs.h \
-                    stats.c stats.h \
+                    stats_prefix.c \
+                    stats_prefix.h \
                     thread.c \
                     thread.h \
                     mc_util.c \
